Wrapping text into container widget
Container can be wrapped around our text here, so open and close parentheses after text and container takes a child named argument which now is the text,

so that is now the widget that's wrapped into container. Container by default is pretty invisible, doesn't have anything that you would see on the screen.
we can set a width and you can set this to double.infinity, .infinity gives you basically a width that ensures that the container takes as much size, as much width as it can get, so by default the full width of the device.