Intro to maps (dictionary)
question here now has to become a bit more complex because now a question should not just be a string, the question itself, It should also contain information about the available answers.

We could create a new class which has all these features we need and use that class to create objects

We create a map with curly braces, which is a collection of key value pairs. like a dictionary in python.