Mapping list to widgets (IMP)
It would be nice if we could create these answer widgets dynamically based on the current question.

To do that, we use maps as list.  This is an anonymous function. answer is every key in the list for the selected question and Creating a new answer widget, and forwarding the answer text to it.

These three dots do here is they take a list which is exactly what we have here and they pull all the values in the list out of it and add them to the surrounding list as individual values, so that we don't add a list to a list
but the values of a list to a list, having only one list without a nested list.